placed input area at the bottom
